DataStreams: flatten Configuration module into single file
Inline `DataStreams::Configuration::Settings` into
`lib/datadog/data_streams/configuration.rb` and remove the empty
`Configuration` wrapper plus `configuration/settings.rb`. Mirror the
RBS layout in `sig/`.

This matches the single-file pattern already used by
`lib/datadog/open_feature/configuration.rb` and reduces boilerplate.
No behavior change.
