+++ sig/datadog/data_streams/configuration.rbs
@@ -1,6 +1,9 @@
 module Datadog
   module DataStreams
     module Configuration
+      module Settings
+        def self.extended: (Module base) -> void
+      end
     end
   end
 end
