--- sig/datadog/data_streams/configuration/settings.rbs
@@ -1,10 +1,0 @@
-module Datadog
-  module DataStreams
-    module Configuration
-      module Settings
-        def self.extended: (Module base) -> void
-      end
-    end
-  end
-end
-
